requirements.txt
pandas>=1.5.0,<3.0.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<1.0.0
langchain>=0.1.0,<1.0.0
langchain-community>=0.0.10,<1.0.0
pyyaml>=6.0,<7.0
tqdm>=4.64.0,<5.0.0
sentence-transformers>=2.2.0,<3.0.0
faiss-cpu>=1.7.0,<2.0.0
ollama>=0.1.0,<1.0.0
interruptingcow>=0.8,<1.0
transformers>=4.30.0,<5.0.0
torch>=1.13.0,<3.0.0